udelay: Add cast for `tv_nsec` to always meet format string
There are libc's around where `tv_nsec` isn't a `long`, e.g. glibc
on x32. Casting to long should always be safe, as `tv_nsec` is de-
fined to be within [0, 999'999'999] (on this code path even within
[0, 100]).

Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/400

+++ udelay.c
@@ -58,7 +58,7 @@ static int clock_check_res(void)
 	if (!clock_getres(clock_id, &res)) {
 		if (res.tv_sec == 0 && res.tv_nsec <= 100) {
 			msg_pinfo("Using clock_gettime for delay loops (clk_id: %d, resolution: %ldns).\n",
-				  (int)clock_id, res.tv_nsec);
+				  (int)clock_id, (long)res.tv_nsec);
 			use_clock_gettime = true;
 			return 1;
 		}
